+++ RELEASENOTES.md
@@ -31,15 +31,15 @@ These are changes that have backwards-compatible solutions present and that comp
 **MAJOR CHANGES**
 
 - Datepicker updates
-    - `className` now is added to the node with `slds-datepicker` class within the dialog. Use `triggerClassName` for the outer wrapping `div`. *[BRAKING CHANGE]*
+    - `className` now is added to the node with `slds-datepicker` class within the dialog. Use `triggerClassName` for the outer wrapping `div`. *[BREAKING CHANGE]*
     - `onChange` (formerly onDateChange) now provides callback with event and a data object in the shape: `{date: [Date object], formattedDate: [string], timezoneOffset: [number]}`
     - Many new props and features: calendar alignment, assistive text props, accepts custom input, supports ISO weekday (Monday first), new props to allow calendar to be controlled (onClose, onOpen, onRequestClose, onRequestOpen), `portalMount` prop for testing React root nodes, more alignment with other menu/dialog components
     - Additional accessiblity and testing
     - See [Date Picker documentation](https://react.lightningdesignsystem.com/components/datepickers/) for fullset of changes
-- `id` on GlobalNav Link move to list item instead of `a`
+- `id` on GlobalNav Link move to list item instead of `a` *[BREAKING CHANGE]*
 - In Button, `aria-*` props now use hyphen instead of camelCase. Use `aria-expanded` and not `ariaExpanded`. *[BREAKING CHANGE]*
 
-**MIONR CHANGES**
+**MINOR CHANGES**
 
 - Picklist bugfix where the value might be available before the options object is. This would make the active item index be incorrect.
 - Fix sort arrow bug in DataTable. Only one column can be actively sorted at a time. Column heading padding has also been updated.